Fix camera pitch issue due to glm::angle() returning degrees
This commit resolves a bug where the camera was unable to pitch because
the glm 0.9.5 version of the glm::angle() function returns degrees
instead of radians. Updated the calculations to convert degrees to
radians to ensure correct camera behavior.

For more details, see: [Stack Overflow](https://stackoverflow.com/questions/79054516/fps-camera-system-cant-look-up-or-down)